Add files via upload
The maze_generator is executed using command python maze_generator.py
Creates a .bin file of the size 20 by 20
it will be updated to get the size of the maze so that it is more dynamic
It uses the mazelib library generating maps with Dungeonrooms and growing trees algorithm.
I also added some random loops for adding complexity and opening up possibility
I also cleared up forst two rows randomly as it was generating walls in the first two rows.

In maze_reader.cpp
it reads the .bin file and assign it to a 2d vector.
The vector is iteratied and each 0 location gets a heuristic that is scaled between 0 to 7 which depends upon the goal that will be set later on to be specified in the command line

the file now is first compiled using
g++ maze_reader.cpp -o maze_reader -std=c++11
and then runn using
./maze_reader
the vector is a 4 bit per gird where the LSB is for representing wall for 1 and path for 0
the first 3 significant bits of the 0 or path is replaced with a scaled heuristic. The heuristic we used here is the Manhattan distance that can be used in Greddy algorithms for path finding.
This bit complexity is used in order to use less space.